+++ lib/src/components/form/fields/select.dart
@@ -53,6 +53,12 @@ class ShadSelectFormField<T> extends ShadFormBuilderField<T> {
     Widget? footer,
     bool allowDeselection = false,
     bool closeOnSelect = true,
+
+    /// {@macro ShadSelect.itemCount}
+    int? itemCount,
+
+    /// {@macro ShadSelect.shrinkWrap}
+    bool? shrinkWrap,
   }) : super(
           decorationBuilder: (context) =>
               (ShadTheme.of(context).selectTheme.decoration ??
@@ -88,6 +94,8 @@ class ShadSelectFormField<T> extends ShadFormBuilderField<T> {
               header: header,
               footer: footer,
               closeOnSelect: closeOnSelect,
+              itemCount: itemCount,
+              shrinkWrap: shrinkWrap,
             );
           },
         );
@@ -143,6 +151,12 @@ class ShadSelectFormField<T> extends ShadFormBuilderField<T> {
     Widget? footer,
     bool allowDeselection = false,
     bool closeOnSelect = true,
+
+    /// {@macro ShadSelect.itemCount}
+    int? itemCount,
+
+    /// {@macro ShadSelect.shrinkWrap}
+    bool? shrinkWrap,
   }) : super(
           decorationBuilder: (context) =>
               (ShadTheme.of(context).selectTheme.decoration ??
@@ -185,6 +199,8 @@ class ShadSelectFormField<T> extends ShadFormBuilderField<T> {
               header: header,
               footer: footer,
               closeOnSelect: closeOnSelect,
+              itemCount: itemCount,
+              shrinkWrap: shrinkWrap,
             );
           },
         );
@@ -237,6 +253,12 @@ class ShadSelectFormField<T> extends ShadFormBuilderField<T> {
     Widget? footer,
     bool allowDeselection = false,
     bool closeOnSelect = true,
+
+    /// {@macro ShadSelect.itemCount}
+    int? itemCount,
+
+    /// {@macro ShadSelect.shrinkWrap}
+    bool? shrinkWrap,
   })  : assert(
           variant == ShadSelectVariant.primary || onSearchChanged != null,
           'onSearchChanged must be provided when variant is search',
@@ -287,6 +309,8 @@ class ShadSelectFormField<T> extends ShadFormBuilderField<T> {
               footer: footer,
               allowDeselection: allowDeselection,
               closeOnSelect: closeOnSelect,
+              itemCount: itemCount,
+              shrinkWrap: shrinkWrap,
             );
           },
         );